engine: extract perft and split UCI modules
