--- include/uci.hpp
@@ -1,168 +1,0 @@
-#pragma once
-
-#include <cstdint>
-#include <optional>
-#include <ostream>
-#include <string>
-#include <string_view>
-#include <variant>
-#include <vector>
-
-#include "defs.hpp"
-#include "move.hpp"
-
-class Board;
-struct RootLine;
-
-// UCI Protocol: https://www.wbec-ridderkerk.nl/html/UCIProtocol.html
-
-namespace uci {
-
-struct SpinOption {
-    int  value;
-    int  def_value;
-    int  min_value;
-    int  max_value;
-    void set(std::string value_str);
-};
-
-struct CheckOption {
-    bool value;
-    bool def_value;
-    void set(std::string val_str);
-};
-
-struct ButtonOption {};
-
-enum class ConfigOption { Hash, Threads, Debug, ClearHash };
-
-struct Config {
-    SpinOption hash = {
-        .value     = DEFAULT_HASH,
-        .def_value = DEFAULT_HASH,
-        .min_value = 1,
-        .max_value = 2048,
-    };
-
-    SpinOption threads = {
-        .value     = DEFAULT_THREADS,
-        .def_value = DEFAULT_THREADS,
-        .min_value = 1,
-        .max_value = 64,
-    };
-
-    CheckOption debug = {
-        .value     = false,
-        .def_value = false,
-    };
-
-    ButtonOption clear_hash;
-
-    ConfigOption set_option(const std::string& name, const std::string& value);
-};
-
-struct EmptyCommand {};
-struct UciCommand {};
-struct DebugCommand {
-    std::string value;
-};
-struct IsReadyCommand {};
-struct SetOptionCommand {
-    std::string name;
-    std::string value;
-    bool        has_value{false};
-};
-struct NewGameCommand {};
-struct PositionCommand {
-    enum class Source { Invalid, Startpos, Fen };
-
-    Source source{Source::Invalid};
-    // Populated only when source == Source::Fen; Startpos is resolved by Engine.
-    std::string              fen;
-    std::vector<std::string> moves;
-};
-struct GoLimits {
-    std::optional<int> depth;
-    std::optional<int> movetime;
-    std::optional<int> nodes;
-    std::optional<int> wtime;
-    std::optional<int> btime;
-    std::optional<int> winc;
-    std::optional<int> binc;
-    std::optional<int> movestogo;
-
-    bool                     searchmoves{false};
-    bool                     ponder{false};
-    bool                     infinite{false};
-    std::optional<int>       mate;
-    std::vector<std::string> unknown_tokens;
-};
-struct GoCommand {
-    GoLimits limits;
-};
-struct StopCommand {};
-struct PonderHitCommand {};
-struct RegisterCommand {};
-struct QuitCommand {};
-struct ExitCommand {};
-struct ConsoleCommand {
-    enum class Name { Help, Board, Eval, Move, Moves, Perft };
-
-    Name        name;
-    std::string arguments;
-};
-struct UnknownCommand {
-    std::string token;
-};
-
-using Command = std::variant<EmptyCommand,
-                             UciCommand,
-                             DebugCommand,
-                             IsReadyCommand,
-                             SetOptionCommand,
-                             NewGameCommand,
-                             PositionCommand,
-                             GoCommand,
-                             StopCommand,
-                             PonderHitCommand,
-                             RegisterCommand,
-                             QuitCommand,
-                             ExitCommand,
-                             ConsoleCommand,
-                             UnknownCommand>;
-
-Command parse_command(std::string_view line);
-
-std::string format_uci_move(Move move);
-std::string format_option(std::string_view name, const SpinOption& opt);
-std::string format_option(std::string_view name, const CheckOption& opt);
-std::string format_option(std::string_view name, const ButtonOption& opt);
-std::string format_identification(const uci::Config& config);
-std::string format_ready();
-std::string format_bestmove(Move move);
-std::string format_info_string(std::string_view str);
-
-// Stateless writer for UCI stdout lines and local debug-console stderr output.
-class Writer {
-public:
-    explicit Writer(std::ostream& out, std::ostream& err) : out(out), err(err) {}
-
-    void help() const;
-    void identify(const uci::Config& config) const;
-    void ready() const;
-    void bestmove(Move move) const;
-    void search_info(const RootLine& line,
-                     const Board&    root_board,
-                     uint64_t        nodes,
-                     Milliseconds    time) const;
-    void info(const std::string& str) const;
-
-    template <typename T>
-    void debug(T&& obj) const;
-
-private:
-    std::ostream& out;
-    std::ostream& err;
-};
-
-} // namespace uci
