+++ tests/search_worker.test.cpp
@@ -8,7 +8,7 @@
 #include "test_util.hpp"
 #include "thread_test_access.hpp"
 #include "threading.hpp"
-#include "uci.hpp"
+#include "uci_writer.hpp"
 
 namespace {
